requirements.txt
folium==0.17.0
matplotlib==3.9.1.post1
pyserial==3.5
PySide6==6.7.2
PySide6==6.7.2
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.2
python-dotenv==1.0.1
